--- HTTP2PushModule/HTTP2PushModule/PushSection.xml
@@ -1,8 +1,0 @@
-﻿<!-- usage-->
-<PushGroupSection>
-  <PushGroup name="group1">
-    <PushElement triggers="true" url="~/main.html"></PushElement>
-    <PushElement url="~/app.js"></PushElement>
-    <PushElement url="~/dependency.js"></PushElement>
-  </PushGroup>
-</PushGroupSection>
